requirements.txt
google-api-python-client==2.108.0
pandas==2.1.3
python-dotenv==1.0.0
jupyter==1.0.0
jupyterlab==4.0.8
notebook==7.0.6
apache-airflow==2.7.2
pytest==7.4.3
black==23.11.0
pyarrow==14.0.1
